# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"ProductsByCategories","description":"API documentation for the app","version":"1.0.0"},"servers":[{"url":"http:\/\/localhost","description":"L5 Swagger OpenApi dynamic host server"}],"paths":{"\/api":{},"\/api\/categories":{"get":{"tags":["Categories"],"summary":"List Categories","description":"Returns an array of categories","operationId":"listCategories","parameters":[{"name":"page","in":"query","description":"Current page number in the pagination (starts at 1)","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successfully got categories","content":{"application\/json":{"schema":{"type":"array","items":{"$ref":"#\/components\/schemas\/Category"}}}}}}}},"\/api\/products":{"get":{"tags":["Products"],"summary":"List products","description":"Returns an array of products","operationId":"listProducts","parameters":[{"name":"page","in":"query","description":"Current page number in the pagination (starts at 1)","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successfully got products","content":{"application\/json":{"schema":{"properties":{"data":{"description":"List products","type":"array","items":{"$ref":"#\/components\/schemas\/Product"}},"current_page":{"description":"List products","type":"integer"},"last_page":{"description":"List products","type":"integer"},"total":{"description":"List products","type":"integer"}},"type":"object"}}}}}},"post":{"tags":["Products"],"summary":"Create new product","description":"Returns new product","operationId":"createProduct","requestBody":{"content":{"multipart\/form-data":{"schema":{"required":["name","category_id"],"properties":{"name":{"description":"Create new product","type":"string"},"category_id":{"description":"Create new product","type":"integer"},"image":{"description":"Create new product","type":"file"}},"type":"object"}}}},"responses":{"200":{"description":"Successfully created a product","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/Product"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"name":{"description":"Create new product","type":"string"},"category_id":{"description":"Create new product","type":"string"},"image":{"description":"Create new product","type":"string"}},"type":"object"}}}},"403":{"description":"Forbidden","content":{"application\/json":{"schema":{"properties":{"message":{"description":"Create new product","type":"string"}},"type":"object"}}}}},"security":[{"passport":[]}]}},"\/api\/products\/{id}":{"put":{"tags":["Products"],"summary":"Update a product","description":"Updates the product specified in the path and returns the updated product","operationId":"updateAProduct","parameters":[{"name":"id","in":"path","description":"Id of the product","required":true,"schema":{"type":"integer"}}],"requestBody":{"content":{"multipart\/form-data":{"schema":{"properties":{"name":{"description":"Update a product","type":"string"},"category_id":{"description":"Update a product","type":"integer"},"image":{"description":"Update a product","type":"file"}},"type":"object"}}}},"responses":{"200":{"description":"Successfully updated a product","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/Product"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"name":{"description":"Update a product","type":"string"},"category_id":{"description":"Update a product","type":"string"},"image":{"description":"Update a product","type":"string"}},"type":"object"}}}},"403":{"description":"Forbidden","content":{"application\/json":{"schema":{"properties":{"message":{"description":"Update a product","type":"string"}},"type":"object"}}}}},"security":[{"passport":[]}]},"delete":{"tags":["Products"],"summary":"Delete a product","description":"Deletes the product specified in the path","operationId":"deleteAProduct","parameters":[{"name":"id","in":"path","description":"Id of the product","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Successfully deleted a product"},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"message":{"description":"Delete a product","type":"string"}},"type":"object"}}}}},"security":[{"passport":[]}]}}},"components":{"schemas":{"Category":{"properties":{"name":{"type":"string"}},"type":"object"},"Product":{"properties":{"name":{"type":"string"}},"type":"object"}},"securitySchemes":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","flow":"password","tokenUrl":"http:\/\/localhost\/oauth\/token","scopes":[],"flows":{"password":{"authorizationUrl":"http:\/\/localhost\/oauth\/authorize","tokenUrl":"http:\/\/localhost\/oauth\/token","refreshUrl":"http:\/\/localhost\/token\/refresh","scopes":[]}}}}}}